gee/pipeline/requirements.txt
earthengine-api>=1.1.0
google-cloud-bigquery>=3.12.0
google-cloud-storage>=2.10.0
functions-framework>=3.0.0
google-auth>=2.23.0
